Move ClangLanguageService to its own module
