+++ Sources/CMakeLists.txt
@@ -5,6 +5,7 @@ add_subdirectory(BuildServerProtocol)
 add_subdirectory(BuildServerIntegration)
 add_subdirectory(CAtomics)
 add_subdirectory(CCompletionScoring)
+add_subdirectory(ClangLanguageService)
 add_subdirectory(CompletionScoring)
 add_subdirectory(Csourcekitd)
 add_subdirectory(Diagnose)
